Twiddle Freeze/Thaw logic some
Per-line Freeze/Thaw was causing delay in the last line appearing.  Also
the ansi-color-test wouldn't scroll because it was frozen for the whole
thing.

This commit overrides Thaw() so that it always Refresh()es and
ScrollsIfAppropriate()

It also (hopefully temporarily) removes the Freeze/Thaw pair around each
line in display().