(Finally) make prefs think in actual boolean; make no-ANSI work
The string-only config scheme writes True and False as the strings
"True" and "False".  Mostly I was remembering to check like:

    if (prefs.get('my_pref') == 'True'):

...but not always and that's just dumb to have to do over and over.  So
now we do an ugly hack in two places to turn the strings "True" and
"False" into True and False when we get them from the config file.

Hopefully no actual value will ever want to be the string "True" or
"False".

Also as a side effect of this, turning off "use ANSI" would make no
output happen at all.  Now it does, but it's still full of ANSI
gibberish;  left to decide whether we should still filter out the escape
codes and then just ignore them.