Initial release
This is in terrible shape following a couple years of lying fallow.
Nonetheless, it fires up now, and connects, and MCP at least somewhat
works, and so forth.  It needs a lot of work but it runs.  If you happen
to check it out, open up issues on the github page if you find things.

This also makes an attempt to use Github Actions to build Windows,
MacOS, and Linux binaries, but we shall see how that turns out.